backend/requirements.txt
# # FastAPI and server
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6

# # Database
# SQLAlchemy==2.0.23
# PyMySQL==1.1.0
# cryptography==41.0.7
# alembic==1.12.1

# # Authentication and Security
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-dotenv==1.0.0

# # Validation
# #pydantic==2.5.0
# pydantic
# pydantic-settings
# email-validator==2.1.0

# # CORS
# # python-cors==1.0.0

# # Testing
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx==0.25.2

# # Utilities
# python-dateutil==2.8.2
# pytz==2023.3

alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==41.0.7
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyMySQL==1.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
mysql-connector-python
